MDEV-26278 Add functionality to eliminate derived tables from the query
Elimination of unnecessary tables from SQL queries is already present
in MariaDB. But it only works for regular tables and not for derived ones.

Imagine we have a view:
  CREATE VIEW v2c AS
  SELECT t11.a AS a, max(t12.col1) AS b
  FROM t11 LEFT JOIN t12 ON t12.pk=t11.b
  GROUP BY t11.a

Due to "GROUP BY t11.a" the values of "a" are unique,
and this fact can be treated as like table "v2c" has a unique key
on field "a".

Suppose we have a SQL query:
  SELECT t1.* FROM t1 LEFT JOIN v2c ON t1.a=v2c.a

Since "v2c.a" is unique, then:
1. "v2c" is functionally dependent on t1.
   This means every record of "t1" will be either joined with
   a single record from "v2c" or NULL-complemented.
2. No fields of "v2c" are present on the SELECT list

These two facts allow to completely exclude (eliminate)
the derived table "v2c" from the query.